refactor: :recycle: cr fixes
