+++ packages/workspace-fusion-modules/src/lib/modules/fusionBookmarks/index.ts
@@ -1,0 +1,1 @@
+export * from './module';
